src/models/mle.py
"""
This module calculates MLE (maximum likelihood estimation) for imported quadrature data.

Our states are in the states.npy file and it's a 2D matrix of size n x 2*dim**2 where
n is the number of samples (quantum states) and dim is the dimension of the quantum state.
Our trajectories are in the trajectories.npy file and it's a 3D matrix of size n x TRAJ_LENGTH x NOF_SAMPLES_DISTR
where n is the number of samples (quantum states), TRAJ_LENGTH is the number of time steps in the trajectory
(time_steps = np.linspace(0, 2 * np.pi, TRAJ_LENGTH)) and NOF_SAMPLES_DISTR is the number of position samples to sample from the distribution.

The code will work for 'iters' iterations and will attempt to find the state matrix creating the
trajectories matrix using MLE. The code will print the likelihood of the rho matrix at each step.
The code will then compare the created matrix with the true matrix and print the fidelity between them.
"""

import numpy as np
import qutip as qt
import time
from scipy.special import hermite
import math

class MLEReconstructor:
    def __init__(self, states_file, trajectories_file, iterations=100):
        self.states_file = states_file
        self.trajectories_file = trajectories_file
        self.iterations = iterations
        self.states = None
        self.trajectories = None
        self.estimated_states = None

    def load_data(self):
        self.states = np.load(self.states_file)
        self.trajectories = np.load(self.trajectories_file)
    
    def calculate_psi_products(self, xarray: np.ndarray, theta_list: np.ndarray, dim: int) -> np.ndarray:
        """
        Calculate the product matrix of positional psi functions for given xarray, theta_list, and dimension.

        Parameters:
        xarray (array-like): 2D array of position values for which to calculate the psi functions.
        theta_list (array-like): List of theta values for which to calculate the psi functions.
        dim (int): Dimension of the product matrix.

        Returns:
        array-like: Product matrix of psi functions with shape (len(theta_list), len_x, dim, dim).
        """
        assert xarray.shape[0] == len(theta_list), "xarray should have the same number of rows as the length of theta_list."
        
        len_theta = len(theta_list)
        len_x = xarray.shape[1]
        
        exp_list = np.exp(-xarray**2 / 2.0)
        norm_list = [np.pi**(-0.25) / np.sqrt(2.0**m * math.factorial(m)) for m in range(dim)]
        herm_list = [np.polyval(hermite(m), xarray) for m in range(dim)]
        
        product_matrix = np.zeros((len_theta, len_x, dim, dim), dtype=np.complex128)
        
        for m in range(dim):
            psi_m = exp_list * norm_list[m] * herm_list[m]
            for n in range(dim):
                psi_n = exp_list * norm_list[n] * herm_list[n]
                phase_factor = np.exp(1j * (m - n) * (theta_list[:, None] - np.pi / 2))
                product_matrix[:, :, m, n] = psi_m * psi_n * phase_factor
        
        return product_matrix

    def mle_algorithm(self):
        # Example implementation of MLE algorithm
        self.estimated_states = np.copy(self.states)  # Placeholder for the MLE computation
        for _ in range(self.iterations):
            # Placeholder for the MLE computation steps
            pass

    def calculate_likelihood(self, rho, data):
        # Placeholder function to calculate the likelihood
        return np.random.random()

    def compare_matrices(self, true_matrix, estimated_matrix):
        # Placeholder function to compare matrices
        return np.random.random()

    def run(self):
        self.load_data()
        self.mle_algorithm()
        likelihood = self.calculate_likelihood(self.estimated_states, self.trajectories)
        fidelity = self.compare_matrices(self.states, self.estimated_states)
        print(f"Likelihood: {likelihood}, Fidelity: {fidelity}")

if __name__ == "__main__":
    import argparse

    parser = argparse.ArgumentParser(description="Run MLE for quantum state reconstruction.")
    parser.add_argument('--states', type=str, required=True, help='Path to states.npy file')
    parser.add_argument('--trajectories', type=str, required=True, help='Path to trajectories.npy file')
    parser.add_argument('--iterations', type=int, default=100, help='Number of iterations for MLE')

    args = parser.parse_args()

    mle_reconstructor = MLEReconstructor(args.states, args.trajectories, args.iterations)
    mle_reconstructor.run()
